Add MainModule to generalize the CLI api
